requirements.txt
flask
pillow
numpy
opencv-python
torch
transformers
diffusers
flask
trimesh
pygltflib
numpy
werkzeug